requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==1.16.0
fastapi==0.104.1
fonttools==4.57.0
h11==0.16.0
idna==3.10
joblib==1.5.0
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
numba==0.61.2
numpy==1.26.4
packaging==25.0
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.2.1
portalocker==3.1.1
psutil==7.0.0
pyasn1==0.6.1
pycryptodome==3.22.0
pycryptodomex==3.22.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.12.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytz==2024.1
pywin32==310
pyzipper==0.3.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.2
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
starlette==0.27.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
